Protecting Video Routes

+++ controller/videoController.js
@@ -6,7 +6,7 @@ export const home = async (req, res) => {
     const videos = await Video.find({}).sort({ _id: -1 });
     res.render("home", { pageTitle: "Home", videos });
   } catch (error) {
-    console.log(error);
+    // console.log(error);
     res.render("home", { pageTitle: "Home", videos: [] });
   }
 };
@@ -24,7 +24,7 @@ export const search = async (req, res) => {
       title: { $regex: searchingBy, $options: "i" },
     });
   } catch (error) {
-    console.log(error);
+    // console.log(error);
   }
   res.render("search", { pageTitle: "Search", searchingBy, videos });
 };
@@ -45,7 +45,7 @@ export const postUpload = async (req, res) => {
     description,
     creator: req.user.id,
   });
-  console.log(newVideo);
+  // console.log(newVideo);
   req.user.videos.push(newVideo.id);
   req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
@@ -71,11 +71,14 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
-  res.render("editVideo", { pageTitle: "editVideo" });
 };
 
 export const postEditVideo = async (req, res) => {
@@ -97,9 +100,14 @@ export const deleteVideo = async (req, res) => {
   } = req;
 
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
-    console.log(error);
+    // console.log(error);
   }
   res.redirect(routes.home);
 };
